Oprydning
Jeg har delt pakkerne op i three layer architecture
